feat(structure): extract view_snapshot type with region/observer metadata
Move tile-render types (sprite_screen_bounds, tint_sprite_record,
tile_render_info, draw_points_cache_t) from map.h into a new
view_snapshot.h, guarded by #if defined(TILES).

Introduce view_snapshot as the top-level wrapper with self-describing
metadata — origin, observer_z, size, observer (character_id), and a
monotonic generation counter — populated at dirty-gated cache rebuild.

Drop the remaining unconditional live reads in draw_terrain,
draw_furniture, and draw_trap: the normal visible path now goes
straight to the cached content; live reads are pushed down into the
override/memory branches only.

Document known client-server architecture gaps (L1-L2 vs L3 boundary,
mutable consumption, missing L3 data, serialization, etc.) inline as
a single source of truth for future work.
